// src/interfaces/interfaces.ts
export interface Scheda {
  id: number
  title: string
  image: string
  explicit: boolean
}

export interface Schema {
  id: number
  artist: string
  song: string
  img: string
}

export interface Fetch {
  id: number
  readable: boolean
  title: string
  title_short: string
  title_version: string
  isrc: string
  link: string
  duration: number
  rank: number
  explicit_lyrics: boolean
  explicit_content_lyrics: number
  explicit_content_cover: number
  preview: string
  md5_image: string
  artist: Artist
  album: Album
  type: string
}

export interface Album {
  id: number
  title: string
  cover: string
  cover_small: string
  cover_medium: string
  cover_big: string
  cover_xl: string
  md5_image: string
  tracklist: string
  type: string
}

export interface Artist {
  id: number
  name: string
  link: string
  picture: string
  picture_small: string
  picture_medium: string
  picture_big: string
  picture_xl: string
  tracklist: string
  type: string
}

export interface player {
  currentSong: Fetch | null
  isPlaying: boolean
  volume: number
}